Overall Code Improvements. (check description)
In this commit me and my teammates worked on handling errors and invalid inputs from the user or any sort of malfunction that may cause to crash the program. We also worked on the naming by unifying the names of variables through the same code while following the instructions of naming variables and functions in C++. All to allow the user experience and the code reader to be more clear and concise.